third_party/blink/web_tests/external/wpt/webmcp/idlharness.https.window-expected.txt

This is a testharness.js-based test.
[FAIL] ModelContext interface: existence and properties of interface object
  assert_equals: prototype of self's property "ModelContext" is not Function.prototype expected function "function () { [native code] }" but got function "function EventTarget() { [native code] }"
[FAIL] ModelContext interface: existence and properties of interface prototype object
  assert_equals: prototype of ModelContext.prototype is not Object.prototype expected object "[object Object]" but got object "[object EventTarget]"
[FAIL] ModelContextClient interface: existence and properties of interface object
  assert_own_property: self does not have own property "ModelContextClient" expected property "ModelContextClient" missing
[FAIL] ModelContextClient interface object length
  assert_own_property: self does not have own property "ModelContextClient" expected property "ModelContextClient" missing
[FAIL] ModelContextClient interface object name
  assert_own_property: self does not have own property "ModelContextClient" expected property "ModelContextClient" missing
[FAIL] ModelContextClient interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "ModelContextClient" expected property "ModelContextClient" missing
[FAIL] ModelContextClient interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "ModelContextClient" expected property "ModelContextClient" missing
[FAIL] ModelContextClient interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "ModelContextClient" expected property "ModelContextClient" missing
[FAIL] ModelContextClient interface: operation requestUserInteraction(UserInteractionCallback)
  assert_own_property: self does not have own property "ModelContextClient" expected property "ModelContextClient" missing
Harness: the test ran to completion.
